chore: Moves date utils to utils folder
