+++ superset-frontend/src/SqlLab/actions/sqlLab.js
@@ -23,7 +23,7 @@ import invert from 'lodash/invert';
 import mapKeys from 'lodash/mapKeys';
 import { isFeatureEnabled, FeatureFlag } from 'src/featureFlags';
 
-import { now } from 'src/modules/dates';
+import { now } from 'src/utils/dates';
 import {
   addDangerToast as addDangerToastAction,
   addInfoToast as addInfoToastAction,
