+++ superset-frontend/src/components/Chart/chartReducer.ts
@@ -22,7 +22,7 @@ import { HYDRATE_DASHBOARD } from 'src/dashboard/actions/hydrate';
 import { DatasourcesAction } from 'src/dashboard/actions/datasources';
 import { ChartState } from 'src/explore/types';
 import { getFormDataFromControls } from 'src/explore/controlUtils';
-import { now } from 'src/modules/dates';
+import { now } from 'src/utils/dates';
 import * as actions from './chartAction';
 
 export const chart: ChartState = {
