+++ superset-frontend/src/dashboard/actions/sliceEntities.js
@@ -21,7 +21,6 @@ import { t, SupersetClient } from '@superset-ui/core';
 import rison from 'rison';
 
 import { addDangerToast } from 'src/components/MessageToasts/actions';
-import { getDatasourceParameter } from 'src/modules/utils';
 import { getClientErrorObject } from 'src/utils/getClientErrorObject';
 
 export const SET_ALL_SLICES = 'SET_ALL_SLICES';
@@ -37,6 +36,10 @@ export function fetchAllSlicesStarted() {
 export const FETCH_ALL_SLICES_FAILED = 'FETCH_ALL_SLICES_FAILED';
 export function fetchAllSlicesFailed(error) {
   return { type: FETCH_ALL_SLICES_FAILED, payload: { error } };
+}
+
+export function getDatasourceParameter(datasourceId, datasourceType) {
+  return `${datasourceId}__${datasourceType}`;
 }
 
 const FETCH_SLICES_PAGE_SIZE = 200;
